+++ Surprise_Delivery_App/pubspec.yaml
@@ -45,6 +45,7 @@ dependencies:
   multi_select_flutter: ^4.1.3
   mockito: ^5.4.5
   mocktail: ^1.0.4
+  http: ^1.3.0
 
 dev_dependencies:
   flutter_test:
